# Registry performance