--- a/modules/generator/registry/counter.go
+++ b/modules/generator/registry/counter.go
@@ -20,6 +20,8 @@ type counter struct {
 
 	onAddSeries    func(count uint32) bool
 	onRemoveSeries func(count uint32)
+
+	externalLabels map[string]string
 }
 
 type counterSeries struct {
@@ -31,6 +33,9 @@ type counterSeries struct {
 	// to the desired value.  This avoids Prometheus throwing away the first
 	// value in the series, due to the transition from null -> x.
 	firstSeries *atomic.Bool
+
+	lb         *labels.Builder
+	baseLabels labels.Labels
 }
 
 var (
@@ -46,7 +51,7 @@ func (co *counterSeries) registerSeenSeries() {
 	co.firstSeries.Store(false)
 }
 
-func newCounter(name string, onAddSeries func(uint32) bool, onRemoveSeries func(count uint32)) *counter {
+func newCounter(name string, onAddSeries func(uint32) bool, onRemoveSeries func(count uint32), externalLabels map[string]string) *counter {
 	if onAddSeries == nil {
 		onAddSeries = func(uint32) bool {
 			return true
@@ -61,6 +66,7 @@ func newCounter(name string, onAddSeries func(uint32) bool, onRemoveSeries func(
 		series:         make(map[uint64]*counterSeries),
 		onAddSeries:    onAddSeries,
 		onRemoveSeries: onRemoveSeries,
+		externalLabels: externalLabels,
 	}
 }
 
@@ -98,11 +104,24 @@ func (c *counter) Inc(labelValueCombo *LabelValueCombo, value float64) {
 }
 
 func (c *counter) newSeries(labelValueCombo *LabelValueCombo, value float64) *counterSeries {
+	// base labels
+	baseLabels := make(labels.Labels, 0, 1+len(c.externalLabels))
+
+	// add external labels
+	for name, value := range c.externalLabels {
+		baseLabels = append(baseLabels, labels.Label{Name: name, Value: value})
+	}
+
+	// add metric name
+	baseLabels = append(baseLabels, labels.Label{Name: labels.MetricName, Value: c.metricName})
+
 	return &counterSeries{
 		labels:      labelValueCombo.getLabelPair(),
 		value:       atomic.NewFloat64(value),
 		lastUpdated: atomic.NewInt64(time.Now().UnixMilli()),
 		firstSeries: atomic.NewBool(true),
+		lb:          labels.NewBuilder(baseLabels),
+		baseLabels:  baseLabels,
 	}
 }
 
@@ -121,31 +140,12 @@ func (c *counter) collectMetrics(appender storage.Appender, timeMs int64, extern
 
 	activeSeries = len(c.series)
 
-	labelsCount := 0
-	if activeSeries > 0 && c.series[0] != nil {
-		labelsCount = len(c.series[0].labels.names)
-	}
-
-	// base labels
-	baseLabels := make(labels.Labels, 0, 1+len(externalLabels)+labelsCount)
-
-	// add external labels
-	for name, value := range externalLabels {
-		baseLabels = append(baseLabels, labels.Label{Name: name, Value: value})
-	}
-
-	// add metric name
-	baseLabels = append(baseLabels, labels.Label{Name: labels.MetricName, Value: c.metricName})
-
-	// TODO: avoid allocation on each collection
-	lb := labels.NewBuilder(baseLabels)
-
 	for _, s := range c.series {
-		lb.Reset(baseLabels)
+		s.lb.Reset(s.baseLabels)
 
 		// set series-specific labels
 		for i, name := range s.labels.names {
-			lb.Set(name, s.labels.values[i])
+			s.lb.Set(name, s.labels.values[i])
 		}
 
 		// If we are about to call Append for the first time on a series, we need
@@ -155,14 +155,14 @@ func (c *counter) collectMetrics(appender storage.Appender, timeMs int64, extern
 			// We set the timestamp of the init serie at the end of the previous minute, that way we ensure it ends in a
 			// different aggregation interval to avoid be downsampled.
 			endOfLastMinuteMs := getEndOfLastMinuteMs(timeMs)
-			_, err = appender.Append(0, lb.Labels(), endOfLastMinuteMs, 0)
+			_, err = appender.Append(0, s.lb.Labels(), endOfLastMinuteMs, 0)
 			if err != nil {
 				return
 			}
 			s.registerSeenSeries()
 		}
 
-		_, err = appender.Append(0, lb.Labels(), timeMs, s.value.Load())
+		_, err = appender.Append(0, s.lb.Labels(), timeMs, s.value.Load())
 		if err != nil {
 			return
 		}