--- a/modules/generator/registry/test.go
+++ b/modules/generator/registry/test.go
@@ -109,7 +109,7 @@ func (t *testCounter) name() string {
 	return t.n
 }
 
-func (t *testCounter) collectMetrics(_ storage.Appender, _ int64, _ map[string]string) (activeSeries int, err error) {
+func (t *testCounter) collectMetrics(_ storage.Appender, _ int64) (activeSeries int, err error) {
 	return
 }
 
@@ -156,7 +156,7 @@ func (t *testGauge) name() string {
 	return t.n
 }
 
-func (t *testGauge) collectMetrics(_ storage.Appender, _ int64, _ map[string]string) (activeSeries int, err error) {
+func (t *testGauge) collectMetrics(_ storage.Appender, _ int64) (activeSeries int, err error) {
 	return 0, nil
 }
 
@@ -206,7 +206,7 @@ func withLe(lbls labels.Labels, le float64) labels.Labels {
 	return lb.Labels()
 }
 
-func (t *testHistogram) collectMetrics(_ storage.Appender, _ int64, _ map[string]string) (activeSeries int, err error) {
+func (t *testHistogram) collectMetrics(_ storage.Appender, _ int64) (activeSeries int, err error) {
 	panic("implement me")
 }