… modules/generator/registry/histogram_test.go