**What this PR does**:

Here we make some performance improvements to the `counter` and `gauge` in the registry.  This is a follow-on to #4232 and #4244.  Additionally, we modify the `metric` interface to skip passing of the `externalLabels` now that each implementation has been updated to take those labels at time of `new`.

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Checklist**
- [x] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`