@zalegrala 2024-11-01
```
❯ benchstat main.txt wip.txt
goos: linux
goarch: amd64
pkg: github.com/grafana/tempo/modules/generator
cpu: AMD Ryzen 7 PRO 7840U w/ Radeon 780M Graphics
           │  main.txt   │               wip.txt               │
           │   sec/op    │   sec/op     vs base                │
Collect-16   4.630µ ± 6%   3.790µ ± 9%  -18.14% (p=0.000 n=10)

           │  main.txt   │            wip.txt            │
           │ heap_in_use │ heap_in_use  vs base          │
Collect-16   17.59M ± 1%   17.71M ± 1%  ~ (p=0.109 n=10)

           │   main.txt   │               wip.txt               │
           │     B/op     │     B/op      vs base               │
Collect-16   13.54Ki ± 2%   12.65Ki ± 4%  -6.57% (p=0.000 n=10)

           │  main.txt   │               wip.txt                │
           │  allocs/op  │  allocs/op    vs base                │
Collect-16   30.50 ± 34%   18.00 ± 100%  -40.98% (p=0.037 n=10)
```